added jpa queary

+++ src/main/java/com/example/University/Dto/UniversityStudentInfoDto.java
@@ -9,7 +9,7 @@ import lombok.Setter;
 @AllArgsConstructor
 public class UniversityStudentInfoDto {
 
-  private Long studentId;
+  private Integer studentId;
 
   private String studentName;
